requirements.txt
# Core dependencies
openai
pydantic
python-dotenv
typeguard
